% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{availableSections}
\alias{availableSections}
\title{Available Sections}
\usage{
availableSections(inpfile)
}
\arguments{
\item{inpfile}{full path to EPANET input file}
}
\value{
character vector containing the names of the sections contained in
the EPANET input file (without brackets)
}
\description{
Names of sections available in EPANET input file
}
\seealso{
\code{\link{readEpanetInputFile}}
}
