% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{calculateSpecificEnergyDemand}
\alias{calculateSpecificEnergyDemand}
\title{Calculate Specific Energy Demand}
\usage{
calculateSpecificEnergyDemand(
  waterDemand,
  totalEnergy,
  COLNAMES = list(Q = "Q.m3.per.hour.sum", E = "Kw.hr.per.m3.avg", Eff =
    "Average.Efficiency.avg")
)
}
\arguments{
\item{waterDemand}{minimum constraint for water demand}

\item{totalEnergy}{list element "energyTotal", as retrieved by
\code{runOptimisationStrategy}}

\item{COLNAMES}{list with elements \emph{Q}, \emph{E}, \emph{Eff}, corresponding to
discharge, energy demand and efficiency, respectively. Default:
list(Q = "Q.m3.per.hour.sum", E = "Kw.hr.per.m3.avg", Eff = "Average.Efficiency.avg")}
}
\value{
list with elements \emph{Q}, \emph{E}, \emph{Eff}, holding the column
names of \emph{totalEnergy}, corresponding to discharge, energy demand and
efficiency, respectively. Default: list(Q = "Q.m3.per.hour.sum", E =
"Kw.hr.per.m3.avg", Eff = "Average.Efficiency.avg")
}
\description{
Calculate Specific Energy Demand
}
