% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{calibrateModel}
\alias{calibrateModel}
\title{Calibrate Model}
\usage{
calibrateModel(
  configuration,
  pipeIDs = NULL,
  measured,
  pumpsToCalibrate = NULL,
  parameterName = "Diameter",
  parameterRange,
  showLivePlot = TRUE,
  ...
)
}
\arguments{
\item{configuration}{EPANET parameterisation, e.g. as retrieved by
readEpanetInputFile()}

\item{pipeIDs}{regular expression or name of pipeID(s) to be used for
calibration}

\item{measured}{measurement data for all pumps as data.frame e.g.
data.frame(pumpNames=c("pmpW1", "pmpW2"), measuredQ=c(140,190))}

\item{pumpsToCalibrate}{regular expression or name of pumps to be used for
calibration: e.g. "pmpW1"}

\item{parameterName}{name of ONE EPANET pipe parameters to be calibrated:
e.g. \code{Diameter} or \code{Roughness}}

\item{parameterRange}{min/max range of possible calibration parameter values:
0-1; with parameterRange*cun = newParameterValue, e.g. parameter range =
0.5 -> 50\% reduction of initial value of \code{parameterName} for all
\code{pipeIDs} defined in calibrateModel()}

\item{showLivePlot}{current calibration status is plotted if
showLivePlot=TRUE. Default: TRUE}

\item{\dots}{additional parameters to be passed to
fitnessAdaptedModelConfiguration()}
}
\description{
Calibrate Model
}
