% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{checkReportFileForErrors}
\alias{checkReportFileForErrors}
\title{Check Report File For Errors}
\usage{
checkReportFileForErrors(reportFile)
}
\arguments{
\item{reportFile}{full path to report file}
}
\description{
Check Report File For Errors
}
