% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{createOptimisationResultsTable}
\alias{createOptimisationResultsTable}
\title{Create Optimisation Results Table}
\usage{
createOptimisationResultsTable(
  optimisationStrategies,
  averageWaterDemand,
  currentEnergyDemand,
  onlyBestSolutions = FALSE
)
}
\arguments{
\item{optimisationStrategies}{average daily water demand in m3/h to be satisfied}

\item{averageWaterDemand}{current specific energy demand}

\item{currentEnergyDemand}{should only the best solutions be written to data.frame? Default: FALSE}

\item{onlyBestSolutions}{return only best solutions? (default: FALSE)}
}
\description{
Create Optimisation Results Table
}
