% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{curvesToText}
\alias{curvesToText}
\title{Curves To Text}
\usage{
curvesToText(curves)
}
\arguments{
\item{curves}{data frame representing \code{curves}, with columns \emph{ID},
\emph{X_Value}, \emph{Y_Value}, as returned by
\code{\link{readEpanetInputFile}} in list element \emph{CURVES}}
}
\description{
Curves To Text
}
