% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureReport.R
\name{defaultReportVariables}
\alias{defaultReportVariables}
\title{Default Report Variables}
\usage{
defaultReportVariables()
}
\value{
list of report variable definitions as returned by
\code{\link{reportVariable}}
}
\description{
Default Report Variables
}
