% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{epanetInputFileLines}
\alias{epanetInputFileLines}
\title{Epanet Input File Lines}
\usage{
epanetInputFileLines(inpdat, dbg = FALSE)
}
\arguments{
\item{inpdat}{input data to be saved in EPANET's input file format}

\item{dbg}{if TRUE, debug messages are shown. Default: FALSE}
}
\description{
Epanet Input File Lines
}
