% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{getLinkResults}
\alias{getLinkResults}
\title{Get Link Results}
\usage{
getLinkResults(
  outdat,
  links,
  vars = c("q", "v", "hl", "wq", "sta", "set", "rr", "ff")
)
}
\arguments{
\item{outdat}{output data read from EPANET output file, as returned by
\code{\link{readEpanetOutputFile}}}

\item{links}{names of \code{links} to be included in the returned data frame. You may use
\code{\link{getNamesOfPipes}, \link{getNamesOfPumps}, 
  \link{getNamesOfValves}} in order to get the names of available \code{links}}

\item{vars}{acronyms of variables to be included in the returned data frame.
"Q" = flow, "v" = velocity, "hl" = headloss, "wq" = avg. water quality,
"sta" = status, "set" = setting, "rr" = reaction rate,
"ff" = friction factor}
}
\description{
Get Link Results
}
