% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{getLinkTimeseriesFromOutputData}
\alias{getLinkTimeseriesFromOutputData}
\title{Get Link Timeseries From Output Data}
\usage{
getLinkTimeseriesFromOutputData(outdat)
}
\arguments{
\item{outdat}{data structure read from EPANET output file, as returned by
\code{\link{readEpanetOutputFile}}.}
}
\description{
Get Link Timeseries From Output Data
}
