% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getNamesOfCurves}
\alias{getNamesOfCurves}
\title{Get Names Of Curves}
\usage{
getNamesOfCurves(inpdat, pattern = ".*")
}
\arguments{
\item{inpdat}{imported EPANET file (as retrieved by \code{readEpanetInputFile})}

\item{pattern}{optional filter pattern (default: ".*")}
}
\description{
Get Names Of Curves
}
