% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getNamesOfJunctions}
\alias{getNamesOfJunctions}
\title{Get Names Of Junctions}
\usage{
getNamesOfJunctions(inpdat, pattern = ".*")
}
\arguments{
\item{inpdat}{imported EPANET file (as retrieved by \code{readEpanetInputFile})}

\item{pattern}{optional filter pattern (default: ".*")}
}
\description{
Get Names Of Junctions
}
