% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getNamesOfReservoirs}
\alias{getNamesOfReservoirs}
\title{Get Names Of Reservoirs}
\usage{
getNamesOfReservoirs(inpdat, pattern = ".*")
}
\arguments{
\item{inpdat}{imported EPANET file (as retrieved by \code{readEpanetInputFile})}

\item{pattern}{optional filter pattern (default: ".*")}
}
\description{
Get Names Of Reservoirs
}
