% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getNamesOfValves}
\alias{getNamesOfValves}
\title{Get Names Of Valves}
\usage{
getNamesOfValves(inpdat, pattern = ".*")
}
\arguments{
\item{inpdat}{imported EPANET file (as retrieved by \code{readEpanetInputFile})}

\item{pattern}{optional filter pattern (default: ".*")}
}
\description{
Get Names Of Valves
}
