% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{getNodeTimeseriesFromOutputData}
\alias{getNodeTimeseriesFromOutputData}
\title{Get Node Timeseries From Output Data}
\usage{
getNodeTimeseriesFromOutputData(outdat)
}
\arguments{
\item{outdat}{data structure read from EPANET output file, as returned by
\code{\link{readEpanetOutputFile}}.}
}
\description{
Get Node Timeseries From Output Data
}
