% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{getNumberOfPeriods}
\alias{getNumberOfPeriods}
\title{Number of Simulation Periods}
\usage{
getNumberOfPeriods(configuration)
}
\description{
number of simulation periods, calculated from duration and hydraulic time
step both of which must be given in the \link{TIMES} section of the EPANET
configuration
}
