% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getPipeCoordinates}
\alias{getPipeCoordinates}
\title{Get Pipe Coordinates}
\usage{
getPipeCoordinates(inpdat)
}
\arguments{
\item{inpdat}{imported EPANET file (as retrieved by \code{readEpanetInputFile})}
}
\description{
Get Pipe Coordinates
}
