% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getPumpInfo}
\alias{getPumpInfo}
\title{Get Pump Info}
\usage{
getPumpInfo(inpdat)
}
\arguments{
\item{inpdat}{imported EPANET file (as retrieved by \code{readEpanetInputFile})}
}
\description{
Get Pump Info
}
