% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{getPumpPerformance}
\alias{getPumpPerformance}
\title{Get Pump Performance}
\usage{
getPumpPerformance(inpdata, outdata, pumpnames)
}
\arguments{
\item{inpdata}{data structure read from EPANET input file, as returned by
\code{\link{readEpanetInputFile}}.}

\item{outdata}{data structure read from EPANET output file, as returned by
\code{\link{readEpanetOutputFile}}.}

\item{pumpnames}{vector with names of pumps for which the pump performance should be evaluated}
}
\value{
data frame with columns \emph{Q} (discharge), \emph{H} (head),
\emph{Eff} (efficiency), \emph{specEn} (specific efficiency),
\emph{En} (energy)
}
\description{
Get time series of pump performance from EPANET result using head curves
and efficiency curves as contained in EPANET input file
}
\seealso{
\code{\link{plotPumpPerformance}}
}
