% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{getSection}
\alias{getSection}
\title{Get Section}
\usage{
getSection(inpfile, sectionName)
}
\arguments{
\item{inpfile}{full path to EPANET input file}

\item{sectionName}{name of section to be read, for possible section names see
the documentation of the EPANET Toolkit}
}
\value{
data frame representing the content of the section in the input file.
If possible, column names are read from the section's header line
}
\description{
Get section from EPANET input file
}
