% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{outputFileSize}
\alias{outputFileSize}
\title{Size of Binary Output File}
\usage{
outputFileSize(configuration)
}
\arguments{
\item{configuration}{EPANET \code{configuration}, representing an EPANET
input \code{file}, as returned by \code{\link{readEpanetInputFile}}}
}
\value{
named vector of numeric representing output file size in bytes, kB
(rounded) and MB (rounded), respectively
}
\description{
Size of binary output file in Bytes, kB (rounded), MB (rounded)
}
