% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{plotCalibration}
\alias{plotCalibration}
\title{Plot Calibration}
\usage{
plotCalibration(newRes)
}
\arguments{
\item{newRes}{expects data.frame object "newRes" as input parameter (is automatically
produced by calibrateModel(). Required columns: \emph{Qerror},
\emph{calibrationRunNumber}, \emph{pch}, \emph{colors}, \emph{pumpNames}}
}
\description{
Plot Calibration
}
