% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{plotCurves}
\alias{plotCurves}
\title{Plot Curves}
\usage{
plotCurves(curves, curveNames = unique(curves$ID), ...)
}
\arguments{
\item{curves}{data frame containing pump \code{curves}, as returned by
\code{\link{readEpanetInputFile}} in list element \emph{CURVES}}

\item{curveNames}{names of \code{curves} to be plotted. Default: all names in
column \emph{ID} of \emph{curves}}

\item{\dots}{additional arguments passed to xyplot}
}
\description{
Plot Curves
}
