% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotModel}
\alias{plotModel}
\title{Plot Model}
\usage{
plotModel(
  inpdat,
  pch = 16,
  cex = 0.2,
  xlab = "",
  ylab = "",
  main = .defaultMain(inpdat, 1),
  zoomToPumps = FALSE,
  changedPipeIDs = NULL,
  ...
)
}
\description{
plot EPANET model network. Allows plotting of changed pipe ids, which are
defined as vector in parameter "changedPipeIDs"
}
