% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{plotOptimisationResults}
\alias{plotOptimisationResults}
\title{Plot Optimisation Results}
\usage{
plotOptimisationResults(
  totalEnergy,
  name = "",
  pumpsToReplace = "",
  userConstraints,
  currentOperation,
  ...
)
}
\description{
Plot Optimisation Results
}
