% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{plotPumpPerformance}
\alias{plotPumpPerformance}
\title{Plot Pump Performance}
\usage{
plotPumpPerformance(xCols, yCols, pumpPerformanceTimeSeries)
}
\arguments{
\item{xCols}{vector of columns contained in data.frame retrieved by getPumpPerformance()
to be used as for x axis plotting, e.g. c("step", "Q")}

\item{yCols}{vector of columns contained in data.frame retrieved by getPumpPerformance()
to be used as for y axis plotting, e.g. c("Eff", "specEn", "En")}

\item{pumpPerformanceTimeSeries}{pump performance time series as retrieved by getPumpPerformance()}
}
\description{
Plot time series of discharge, head and pump performance
}
\seealso{
\code{\link{getPumpPerformance}}
}
