% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{readEpanetInputFile}
\alias{readEpanetInputFile}
\title{Read EPANET Input File}
\usage{
readEpanetInputFile(inpfile, dbg = FALSE)
}
\arguments{
\item{inpfile}{full path to EPANET input file}

\item{dbg}{if TRUE, debug messages are shown. Default: FALSE}
}
\value{
list with elements representing the different sections of the EPANET
input file. The names of the list elements correspond to the names of
the sections that were found in the input file. Each list element is a
data frame containing the content of the corresponding section.
}
\description{
Read EPANET Input File
}
\seealso{
\code{\link{availableSections}, \link{readEpanetOutputFile}}
}
