% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{readEpanetOutputFile}
\alias{readEpanetOutputFile}
\title{Read EPANET Output File}
\usage{
readEpanetOutputFile(
  outfile,
  read.prolog = TRUE,
  read.energyUse = TRUE,
  read.dynamicResults = TRUE,
  read.epilog = TRUE
)
}
\arguments{
\item{outfile}{full path to EPANET output file}

\item{read.prolog}{if TRUE, the "Prolog" section is read from the output file and contained
in the output list}

\item{read.energyUse}{if TRUE, the "Energy Use" section is read from the output file and
contained in the output list}

\item{read.dynamicResults}{if TRUE, the "Extended Period" section is read from the output file and
contained in the output list}

\item{read.epilog}{if TRUE, the "Epilog" section is read from the output file and contained
in the output list}
}
\value{
list with elements \emph{prolog} (if \code{read.prolog} = TRUE), \emph{energyUse}
(if \code{read.energyUse} = TRUE), \emph{dynamicResults} (if \code{read.dynamicResults}
= TRUE) and \emph{epilog} (if \code{read.epilog} = TRUE), containing the
different parts of the output file, as described in the documentation of
the EPANET Toolkit.
}
\description{
Read EPANET Output File
}
\seealso{
\code{\link{readEpanetInputFile}}
}
