% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReportFile.R
\name{readResultsFromReportFile}
\alias{readResultsFromReportFile}
\title{Read Results From Report File}
\usage{
readResultsFromReportFile(reportFile, warn = TRUE)
}
\arguments{
\item{reportFile}{full path to report file, generated by EPANET}

\item{warn}{if TRUE, a warning is given if no Node or Link results were found
in the report file.}
}
\value{
list with elements \emph{nodeData} containing a data frame with node
results and \emph{linkData} containing a data frame with link results
}
\description{
reads an EPANET report file and returns a list with node and link data
}
