% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{replaceCurves}
\alias{replaceCurves}
\title{Replace Curves In CURVES Data}
\usage{
replaceCurves(curves, newCurves)
}
\arguments{
\item{curves}{data frame representing \code{curves}, with columns \emph{ID},
\emph{X_Value}, \emph{Y_Value}, as returned by
\code{\link{readEpanetInputFile}} in list element \emph{CURVES}}

\item{newCurves}{new curves data frame representing \code{curves}, with columns \emph{ID},
\emph{X_Value}, \emph{Y_Value},}
}
\description{
Replace \code{curves} in CURVES data
}
