% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{replaceCurveSectionInInputFile}
\alias{replaceCurveSectionInInputFile}
\title{Replace Curve Section In Input File}
\usage{
replaceCurveSectionInInputFile(inpfile, newCurves)
}
\arguments{
\item{inpfile}{full path to EPANET input file}

\item{newCurves}{modified CURVES section. Data frame as returned by
\code{\link{readEpanetInputFile}} in list element \emph{curves}}
}
\value{
vector of character(s) representing the rows of the input file
}
\description{
Replace Curve Section In Input File
}
