% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{replaceFileExtension}
\alias{replaceFileExtension}
\title{Replace File Extension}
\usage{
replaceFileExtension(filename, newExtension)
}
\arguments{
\item{filename}{full path to file of which the extension is to be changed}

\item{newExtension}{new extension to be given to the file, without dot "."}
}
\description{
Replace File Extension
}
