% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{replaceOneCurve}
\alias{replaceOneCurve}
\title{Replace One Curve In CURVES Data}
\usage{
replaceOneCurve(curves, curveName, x, y)
}
\arguments{
\item{curves}{data frame representing \code{curves}, with columns \emph{ID},
\emph{X_Value}, \emph{Y_Value}, as returned by
\code{\link{readEpanetInputFile}} in list element \emph{CURVES}}

\item{curveName}{name of the curve to be replaced}

\item{x}{vector of \code{x} values of the curve}

\item{y}{vector of \code{y} values of the curve}
}
\value{
data frame in which the lines corresponding to the curve named
\emph{curveName} are replaced with \code{x} and \code{y} values given in
\emph{x} and \emph{y}, respectively
}
\description{
Replace One Curve In CURVES Data
}
