% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{reportEnergyUse}
\alias{reportEnergyUse}
\title{Report Energy Use}
\usage{
reportEnergyUse(outdat)
}
\arguments{
\item{outdat}{output data read from EPANET out-file}
}
\description{
Report Energy Use
}
