% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runEpanet}
\alias{runEpanet}
\title{Run EPANET With Given Input File}
\usage{
runEpanet(
  inpfile,
  returnOutput = FALSE,
  epanet.dir = getEpanetInstallationPath(),
  intern = FALSE,
  write.output = TRUE,
  ...,
  dbg = FALSE
)
}
\arguments{
\item{inpfile}{full path to EPANET input file}

\item{returnOutput}{if TRUE, the output is read from the generated output
file and returned}

\item{epanet.dir}{path to EPANET installation directory. Default:
getEpanetInstallationPath()}

\item{intern}{a logical, indicates whether to make the output of the command
an R object.}

\item{write.output}{if TRUE, EPANET will write a binary output file, else not}

\item{\dots}{further arguments passed to \code{\link{readEpanetOutputFile}},
such as: \emph{read.prolog}, \emph{read.energyUse},
\emph{read.dynamicResults}, \emph{read.epilog}, see there.}

\item{dbg}{if TRUE, debug messages are shown. Default: FALSE}
}
\description{
Run EPANET With Given Input File
}
