% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runEpanetConfiguration}
\alias{runEpanetConfiguration}
\title{Run EPANET INP configuration}
\usage{
runEpanetConfiguration(
  inpdat,
  name = "tmpEpanet",
  returnOutput = write.output,
  write.output = TRUE,
  ...,
  dbg = FALSE
)
}
\arguments{
\item{inpdat}{input data as retrieved by \code{\link{readEpanetInputFile}}}

\item{name}{\code{name} of input file to be generated in tempdir()}

\item{returnOutput}{if TRUE, the output is read from the output file (if generated, see
\emph{write.output}) and returned. Default: value of \emph{write.output}}

\item{write.output}{if TRUE, EPANET will write a binary output file, else not}

\item{\dots}{further arguments passed to \code{\link{runEpanet}} and finally to
\code{\link{readEpanetOutputFile}}, such as: \emph{read.prolog},
\emph{read.energyUse}, \emph{read.dynamicResults}, \emph{read.epilog}, see
there.}

\item{dbg}{if TRUE, debug messages are shown. Default: FALSE}
}
\description{
Run EPANET INP configuration
}
