% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runEpanetGUI}
\alias{runEpanetGUI}
\title{Run Epanet GUI}
\usage{
runEpanetGUI(inpfile = "", epanet.dir = getEpanetInstallationPath())
}
\arguments{
\item{inpfile}{path to EPANET input file}

\item{epanet.dir}{EPANET directory (default: \code{getEpanetInstallationPath})}
}
\description{
Run Epanet GUI
}
