% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{runOptimisationStrategy}
\alias{runOptimisationStrategy}
\title{Run Optimisation Strategy}
\usage{
runOptimisationStrategy(
  configuration,
  newCurvesData,
  optimisationStrategy,
  operationSchemes = wellFieldOperationSchemes(getNamesOfPumps(configuration)),
  showLivePlot = FALSE
)
}
\arguments{
\item{operationSchemes}{possible wellfield operation schemes. Default:
wellFieldOperationSchemes(getNamesOfPumps(\code{configuration}))}
}
\value{
list with elements \emph{energyTotal} and \emph{energyPerPump}
}
\description{
Run Optimisation Strategy
}
