% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{setEpanetInstallationPath}
\alias{setEpanetInstallationPath}
\title{Set Epanet Installation Path}
\usage{
setEpanetInstallationPath(epanet.dir)
}
\arguments{
\item{epanet.dir}{full path to MS Access database or ODBC database name}
}
\description{
Set Epanet Installation Path
}
