% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureReport.R
\name{setReportOptions}
\alias{setReportOptions}
\title{Set REPORT Option in EPANET configuration}
\usage{
setReportOptions(
  configuration,
  pagesize = 0,
  file = "",
  status = "NO",
  summary = "YES",
  messages = "YES",
  energy = "NO",
  nodes = "NONE",
  links = "NONE",
  variables = defaultReportVariables()
)
}
\arguments{
\item{configuration}{EPANET \code{configuration}, representing an EPANET
input \code{file}, as returned by \code{\link{readEpanetInputFile}}}

\item{pagesize}{\code{pagesize} sets the number of lines written per page of
the output report. The default is 0, meaning that no line limit per page is
in effect.}

\item{file}{\code{file} supplies the name of a \code{file} to which the
output report will be written. If the \code{file} name contains spaces then
it must be surrounded by double quotes. If not supplied then the Report
\code{file}, as specified in the second parameter of the ENopen (or
ENepanet) function will be used.}

\item{status}{\code{status} determines whether hydraulic \code{status}
\code{messages} are written to the Report \code{file}. If YES is selected
the \code{messages} will identify those network components that change
\code{status} during each time step of the simulation. If FULL is selected,
then convergence information will also be included from each trial of each
hydraulic analysis. This level of detail is only useful for de-bugging
networks that become hydraulically unbalanced. The default is NO.}

\item{summary}{\code{summary} determines whether a \code{summary} table of
number of network components and key analysis options is generated. The
default is YES.}

\item{messages}{\code{messages} determines whether error and warning
\code{messages} generated during a hydraulic/water quality analysis are
written to the Report \code{file}. The default is YES.}

\item{energy}{\code{energy} determines if a table reporting average
\code{energy} usage and cost for each pump is provided. The default is NO.}

\item{nodes}{\code{nodes} identifies which \code{nodes} will be reported on.
You can either list individual node ID labels or use the keywords NONE or
ALL. Additional NODES lines can be used to continue the list. The default
is NONE.}

\item{links}{\code{links} identifies which \code{links} will be reported on.
You can either list individual link ID labels or use the keywords NONE or
ALL. Additional LINKS lines can be used to continue the list. The default
is NONE.}

\item{variables}{list of report \code{variables} as defined by
\code{\link{reportVariable}}, e.g. list(reportVariable(name = "VELOCITY",
above = 3.0, precision = 4), reportVariable(name = "F-FACTOR", precision =
4)). Default: \code{\link{defaultReportVariables}}}
}
\description{
Set REPORT Option in EPANET configuration
}
