% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{setTimeParameter}
\alias{setTimeParameter}
\title{Set EPANET' Time-Related Simulation Parameters}
\usage{
setTimeParameter(
  configuration,
  duration = "",
  hydraulic.timestep = "",
  quality.timestep = "",
  rule.timestep = "",
  pattern.timestep = "",
  pattern.start = "",
  report.timestep = "",
  report.start = "",
  start.clocktime = "",
  statistic = ""
)
}
\description{
Set EPANET' time-related simulation parameters. See EPANET reference for the
meaning of the parameters
}
