% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{setWellFieldOperation}
\alias{setWellFieldOperation}
\title{Set Well Field Operation}
\usage{
setWellFieldOperation(config, operationScheme)
}
\arguments{
\item{config}{EPANET \code{configuration}, representing an EPANET
input \code{file}, as returned by \code{\link{readEpanetInputFile}}}

\item{operationScheme}{one possible wellfield operation scheme
(e.g. wellFieldOperationSchemes(getNamesOfPumps(\code{configuration})))}
}
\value{
config with modified well field operation rules
}
\description{
Set Well Field Operation
}
