% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpanetOutput.R
\name{showProperties}
\alias{showProperties}
\title{Show Properties}
\usage{
showProperties(outdata)
}
\arguments{
\item{outdata}{list structure with EPANET results as retrieved by
\code{\link{readEpanetOutputFile}}}
}
\description{
Show node and link properties available in EPANET output
}
