% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialFunctions.R
\name{wellFieldOperationSchemes}
\alias{wellFieldOperationSchemes}
\title{Well Field Operation Schemes}
\usage{
wellFieldOperationSchemes(
  pumpNames,
  niceLabels = list(searchPattern = "pmp", searchReplacement = "p")
)
}
\arguments{
\item{pumpNames}{Vector of \code{pumpNames} (used in EPANET input file section "PUMPS")}

\item{niceLabels}{Optionally nice labeling. List with elements \emph{searchPattern} and
\emph{searchReplacement}. Default: list(searchPattern = "pmp",
searchReplacement = "p")}
}
\description{
Well Field Operation Schemes
}
