% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCurves.R
\name{writeCurves}
\alias{writeCurves}
\title{Write Curves}
\usage{
writeCurves(
  epanetConfig,
  pumpCurves = NULL,
  efficiencyCurves = NULL,
  drawdownCurves = NULL,
  deleteOldCurves = TRUE,
  dbg = TRUE
)
}
\arguments{
\item{epanetConfig}{EPANET configuration, as retrieved by \code{readEpanetInputFile}}

\item{pumpCurves}{list(data = data.frame(ID, X_Value, Y_Value), label = "")}

\item{efficiencyCurves}{list(data = data.frame(ID, X_Value, Y_Value), label =
"")}

\item{drawdownCurves}{list(data = data.frame(ID, X_Value, Y_Value), label =
"")}

\item{deleteOldCurves}{should all curves in 'epanetConfig' be deleted before
writing?}

\item{dbg}{show debug messages? (default: TRUE)}
}
\description{
Write Curves
}
