% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCurves.R
\name{writeDrawdownCurves}
\alias{writeDrawdownCurves}
\title{Write Drawdown Curves}
\usage{
writeDrawdownCurves(
  epanetConfig,
  DD,
  curveNamePrefix = "dd",
  deleteOldCurves = FALSE,
  dbg = TRUE
)
}
\arguments{
\item{epanetConfig}{EPANET configuration, as retrieved by \code{readEpanetInputFile}}

\item{DD}{data.frame(ID, X_Value, Y_Value)}

\item{curveNamePrefix}{prefix to be used in the curve name (default: "dd")}

\item{deleteOldCurves}{if TRUE all curves in epantetConfig$CURVES will be
deleted before adding new ones (default: FALSE)}

\item{dbg}{show debug messages? (default: TRUE)}
}
\value{
Modified EPANET configuration
}
\description{
Write Drawdown Curves
}
