% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{writeEpanetInputFile}
\alias{writeEpanetInputFile}
\title{Write Epanet Input File}
\usage{
writeEpanetInputFile(inpdat, inpfile, dbg = FALSE)
}
\arguments{
\item{inpdat}{input data to be saved in EPANET's input file format}

\item{inpfile}{full path to input file to be created}

\item{dbg}{if TRUE, debug messages are shown. Default: FALSE}
}
\description{
Write Epanet Input File
}
