% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileAccess.R
\name{writeInputFileWithNewCurveSection}
\alias{writeInputFileWithNewCurveSection}
\title{Write Input File With New Curve Section}
\usage{
writeInputFileWithNewCurveSection(
  inpfile,
  newCurves,
  inpfile.new = sub("(\\\\.[^\\\\.]+)$", "_new\\\\1", inpfile)
)
}
\arguments{
\item{inpfile}{full path to EPANET input file}

\item{newCurves}{modified CURVES section. Data frame as returned by
\code{\link{readEpanetInputFile}} in list element \emph{curves}}

\item{inpfile.new}{full path to modified EPANET input file. (default: \emph{\<inpfile-without-extension_new\>.inp})}
}
\value{
full path to created input file
}
\description{
Write Input File With New Curve Section
}
