% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCurves.R
\name{writePumpCurves}
\alias{writePumpCurves}
\title{Write Pump Curves}
\usage{
writePumpCurves(
  epanetConfig,
  TDH,
  curveNamePrefix = "TDH",
  deleteOldCurves = FALSE,
  dbg = TRUE
)
}
\arguments{
\item{TDH}{data.frame(ID, X_Value, Y_Value)}
}
\value{
Modified EPANET configuration
}
\description{
Write Pump Curves
}
