% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotModel}
\alias{plotModel}
\title{Plot Model}
\usage{
plotModel(
  inpdat,
  pch = 16,
  cex = 0.2,
  xlab = "",
  ylab = "",
  main = defaultMain(inpdat, 1),
  zoomToPumps = FALSE,
  changedPipeIDs = NULL,
  ...
)
}
\arguments{
\item{inpdat}{EPANET input file}

\item{pch}{Either an integer specifying a symbol or a single character to be
used as the default in plotting points. See \link[graphics]{points} for possible
values and their interpretation (default: 16)}

\item{cex}{A numerical value giving the amount by which plotting text and symbols
should be magnified relative to the default. This starts as 1 when a device is
opened, and is reset when the layout is changed, e.g. by setting mfrow.
(default: 0.2)}

\item{xlab}{a title for the x axis (default: "")}

\item{ylab}{a title for the y axis (default: "")}

\item{main}{an overall title for the plot (default: kwb.epanet:::defaultMain(inpdat, 1))}

\item{zoomToPumps}{should plot zoomed to pumps (default: FALSE)}

\item{changedPipeIDs}{optional, vector with changed pipe IDs (default: NULL)}

\item{...}{additional arguments passed to \link[graphics]{plot}}
}
\description{
plot EPANET model network. Allows plotting of changed pipe ids, which are
defined as vector in parameter "changedPipeIDs"
}
