% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runEpanetOnCommandLine}
\alias{runEpanetOnCommandLine}
\title{Run Epanet On Command Line}
\usage{
runEpanetOnCommandLine(
  inpfile,
  epanet.exe,
  intern = FALSE,
  write.output = TRUE,
  dbg = FALSE
)
}
\arguments{
\item{inpfile}{path to EPANET input file}

\item{epanet.exe}{path to EPANET executable}

\item{intern}{a logical, indicates whether to make the output of the command an R object.
(default: FALSE)}

\item{write.output}{if TRUE, EPANET will write a binary output file, else not}

\item{dbg}{if TRUE, debug messages are shown. Default: FALSE}
}
\description{
Run Epanet On Command Line
}
