% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{setTimeParameter}
\alias{setTimeParameter}
\title{Set EPANET' Time-Related Simulation Parameters}
\usage{
setTimeParameter(
  configuration,
  duration = "",
  hydraulic.timestep = "",
  quality.timestep = "",
  rule.timestep = "",
  pattern.timestep = "",
  pattern.start = "",
  report.timestep = "",
  report.start = "",
  start.clocktime = "",
  statistic = ""
)
}
\arguments{
\item{configuration}{EPANET \code{configuration}, representing an EPANET
input \code{file}, as returned by \code{\link{readEpanetInputFile}}}

\item{duration}{(default: "")}

\item{hydraulic.timestep}{(default: "")}

\item{quality.timestep}{(default: "")}

\item{rule.timestep}{(default: "")}

\item{pattern.timestep}{(default: "")}

\item{pattern.start}{(default: "")}

\item{report.timestep}{(default: "")}

\item{report.start}{(default: "")}

\item{start.clocktime}{(default: "")}

\item{statistic}{(default: "")}
}
\value{
return configuration with modified [TIMES] parameterisation
}
\description{
Set EPANET' time-related simulation parameters. See EPANET reference for the
meaning of the parameters
}
