% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCurves.R
\name{writeEfficiencyCurves}
\alias{writeEfficiencyCurves}
\title{Write Efficiency Curves}
\usage{
writeEfficiencyCurves(
  epanetConfig,
  Eff,
  curveNamePrefix = "Eff",
  deleteOldCurves = FALSE,
  dbg = TRUE
)
}
\arguments{
\item{Eff}{data.frame(ID, X_Value, Y_Value)}
}
\value{
Modified EPANET configuration
}
\description{
Write Efficiency Curves
}
