% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCurves.R
\name{writePumpCurves}
\alias{writePumpCurves}
\title{Write Pump Curves}
\usage{
writePumpCurves(
  epanetConfig,
  TDH,
  curveNamePrefix = "TDH",
  deleteOldCurves = FALSE,
  dbg = TRUE
)
}
\arguments{
\item{epanetConfig}{EPANET configuration, as retrieved by \code{readEpanetInputFile}}

\item{TDH}{data.frame(ID, X_Value, Y_Value)}

\item{curveNamePrefix}{(default: "TDH")}

\item{deleteOldCurves}{if TRUE all curves in epantetConfig$CURVES will be
deleted before adding new ones (default: FALSE)}

\item{dbg}{show debug messages? (default: TRUE)}
}
\value{
Modified EPANET configuration
}
\description{
Write Pump Curves
}
